Feature: Mood Buttons (#10)
* Initial branch commit

* Add mood buttons

Added the predefined mood button elements the user can click on to get a track from that buttons defined mood.

* Add Genre Drop Down

Added a drop down element for the user to select a genre to seed the Spotify recommendation.

* Add mood button api

Added the logic and api routes to make the mood buttons work.

* Update styling

* Update styling

The app's look was updated and new favicon files were added. A new handlebars file was added to account for the new mood buttons which use a separate route from the login api.

* Tweak text

Tweaked the handlebars files to support the case where no tracks were returned.